requirements.txt
Flask==3.0.2
selenium==4.18.1
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
webdriver-manager==4.0.1